manage listings implemented, routes secured

+++ app/Http/Controllers/ListingController.php
@@ -52,7 +52,7 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
-        $formFields['user_id']=auth()->id();
+        $formFields['user_id'] = auth()->id();
 
         Listing::create($formFields);
 
@@ -71,6 +71,12 @@ class ListingController extends Controller
     //Update Form
     public function update(Request $request, Listing $listing)
     {
+
+        //Make sure logged in user is owner
+
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorised action');
+        }
 
         $formFields = $request->validate([
             'title' => 'required',
@@ -92,10 +98,25 @@ class ListingController extends Controller
     }
 
     //Delete Listing
-    public function destroy(Listing $listing){
+    public function destroy(Listing $listing)
+    {
+        //Make sure logged in user is owner
+
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorised action');
+        }
+
         $listing->delete();
 
         return redirect('/')->with('message', "Listing deleted successfully!");
+    }
+
+    //Manage Listings
+    public function manage()
+    {
+        return view('listings.manage', [
+            'listings' => auth()->user()->listings()->get()
+        ]);
     }
 
 }
